Added many issues
