+++ app/helpers/mailer_helper.rb
@@ -11,6 +11,24 @@ module MailerHelper
       }
     })
   end
+  # This function will find if there are already reviews present for the current submission,
+  # If the reviews are present then it will mail each reviewer a mail with the link to update the current review.
+
+  def self.mail_assigned_reviewers(team)
+    maps = ResponseMap.where(reviewed_object_id: @participant.assignment.id, reviewee_id: team.id, type: 'ReviewResponseMap')
+    unless maps.nil?
+      maps.each do |map|
+        # Mailing function
+        Mailer.general_email(
+          to: User.find(Participant.find(map.reviewer_id).user_id).email,
+          subject:  "Link to update the review for Assignment '#{@participant.assignment.name}'",
+          cc: User.find_by(@participant.assignment.instructor_id).email,
+          link: "Link: https://expertiza.ncsu.edu/response/new?id=#{map.id}",
+          assignment: @participant.assignment.name
+        ).deliver_now
+      end
+    end
+  end
 
   def self.send_mail_to_all_super_users(super_user, user, subject)
     Mailer.request_user_message ({
